+++ src/test/java/org/jsoup/nodes/ElementTest.java
@@ -2555,9 +2555,9 @@ public class ElementTest {
         Document doc = Jsoup.parse(html);
         assertEquals("<bar>\n <p></p>\n</bar>", doc.body().html());
 
-        html = "<foo>\n  <bar />\n</foo>";
+        html = "<foo>\n  <bar /></foo>";
         doc = Jsoup.parse(html);
-        assertEquals("<foo>\n <bar />\n</foo>", doc.body().html());
+        assertEquals("<foo>\n <bar></bar>\n</foo>", doc.body().html());
     }
 
     @Test void spanInBlockTrims() {